Reflection serializers: test data cleanup
