+++ rd-net/Test.RdFramework/Reflection/TestReflectionSerialization.cs
@@ -11,14 +11,11 @@ namespace Test.RdFramework.Reflection
     [RdExt]
     public sealed class RootModel : RdExtReflectionBindableBase
     {
-      // public NestedModel Nested { get; }
-
       public IViewableProperty<bool> Primitive { get; }
       public IViewableProperty<EmptyOK> EmptyOK { get; }
       public IViewableProperty<FieldsNotNullOk> FieldsNotNullOk { get; }
       public IViewableProperty<FieldsNullableOk> FieldsNullableOk { get; }
       public IViewableProperty<PropertiesNotNullOk> PropertiesNotNullOk { get; }
-      public IViewableProperty<PropertiesNullOk> PropertiesNullOk { get; }
       public IViewableProperty<Animal> PolyProperty { get; }
     }
