--- rd-net/Test.RdFramework/Reflection/data/Bear.cs
@@ -1,7 +1,0 @@
-﻿using JetBrains.Rd.Reflection;
-
-namespace Test.RdFramework.Reflection
-{
-  [RdModel]
-  public sealed class Bear : Mammal {}
-}
