--- rd-net/Test.RdFramework/Reflection/data/CircularDependencyExt2Error.cs
@@ -1,10 +1,0 @@
-﻿using JetBrains.Rd.Reflection;
-
-namespace Test.RdFramework.Reflection
-{
-  [RdExt]
-  public class CircularDependencyExt2Error : RdExtReflectionBindableBase
-  {
-    public CircularDependencyExtError ParentRef { get; }
-  }
-}
