--- rd-net/Test.RdFramework/Reflection/data/CircularDependencyExtError.cs
@@ -1,10 +1,0 @@
-﻿using JetBrains.Rd.Reflection;
-
-namespace Test.RdFramework.Reflection
-{
-  [RdExt]
-  public class CircularDependencyExtError : RdExtReflectionBindableBase
-  {
-    public CircularDependencyExt2Error ChildRef { get; }
-  }
-}
