--- rd-net/Test.RdFramework/Reflection/data/CircularDependencyInModelError.cs
@@ -1,11 +1,0 @@
-﻿using JetBrains.Collections.Viewable;
-using JetBrains.Rd.Reflection;
-
-namespace Test.RdFramework.Reflection
-{
-  [RdExt]
-  public class CircularDependencyInModelError : RdExtReflectionBindableBase
-  {
-    public IViewableProperty<CircularDependencyModelError> ParentRef { get; }
-  }
-}
