--- rd-net/Test.RdFramework/Reflection/data/CircularDependencyModel2Error.cs
@@ -1,11 +1,0 @@
-﻿using JetBrains.Collections.Viewable;
-using JetBrains.Rd.Reflection;
-
-namespace Test.RdFramework.Reflection
-{
-  [RdModel]
-  public class CircularDependencyModel2Error : RdReflectionBindableBase
-  {
-    public IViewableProperty<CircularDependencyModelError> Value;
-  }
-}
