--- rd-net/Test.RdFramework/Reflection/data/CircularDependencyModelError.cs
@@ -1,11 +1,0 @@
-﻿using JetBrains.Collections.Viewable;
-using JetBrains.Rd.Reflection;
-
-namespace Test.RdFramework.Reflection
-{
-  [RdModel]
-  public class CircularDependencyModelError : RdReflectionBindableBase
-  {
-    public IViewableProperty<CircularDependencyModel2Error> Value;
-  }
-}
