+++ rd-net/Test.RdFramework/Reflection/data/CornerCase/CantHaveNonRdError.cs
