+++ rd-net/Test.RdFramework/Reflection/data/CornerCase/NoBaseType.cs
