+++ rd-net/Test.RdFramework/Reflection/data/Example/Animals.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using JetBrains.Annotations;
 using JetBrains.Collections.Viewable;
-using JetBrains.Rd.Base;
 using JetBrains.Rd.Reflection;
 
 namespace Test.RdFramework.Reflection
@@ -12,11 +11,25 @@ namespace Test.RdFramework.Reflection
     /// <summary>
     /// It is possible to have arbitrary data in live models if attribute <see cref="NonSerializedAttribute"/> is specified
     /// </summary>
-    [NonSerialized]
-    [CanBeNull] public string[] arrays;
+    [NonSerialized] [CanBeNull] public string[] arrays;
 
-    public IViewableList<FieldsNotNullOk> LiveList;
+    /// <summary>
+    /// Nested RdModel should work
+    /// </summary>
+    public ModelSample NestedRdModel;
+
+    public IViewableList<ModelSample> LiveList;
     // public IViewableMap<FieldsNotNullOk, int[]> maps { get; set; }
     // public IViewableSet<FieldsNotNullOk> sets;
+  }
+
+  [RdModel]
+  public sealed class Bear : Mammal
+  {
+  }
+
+  [RdModel]
+  public class Mammal : Animal
+  {
   }
 }
