+++ rd-net/Test.RdFramework/Reflection/data/Example/PropertiesNotNullOk.cs
