+++ rd-net/Test.RdFramework/Reflection/data/Example/RootModel.cs
@@ -1,4 +1,5 @@
 ﻿using JetBrains.Collections.Viewable;
+using JetBrains.Rd.Impl;
 using JetBrains.Rd.Reflection;
 
 namespace Test.RdFramework.Reflection
@@ -10,9 +11,18 @@ namespace Test.RdFramework.Reflection
     public NestedModel Nested { get; }
 
     public IViewableProperty<EmptyOK> EmptyOK { get; }
-    public IViewableProperty<FieldsNotNullOk> FieldsNotNullOk { get; }
+    public IViewableProperty<ModelSample> FieldsNotNullOk { get; }
     public IViewableProperty<FieldsNullableOk> FieldsNullableOk { get; }
     public IViewableProperty<PropertiesNotNullOk> PropertiesNotNullOk { get; }
-    public IViewableProperty<PropertiesNullOk> PropertiesNullOk { get; }
+
+    [RdExt]
+    public sealed class NestedModel : RdExtReflectionBindableBase
+    {
+      public RdProperty<string> SomeProperty { get; }
+      public IViewableProperty<string> IPropertyAlsoFine { get; }
+
+      public IViewableProperty<PropertiesNotNullOk> PropModel { get; }
+      public RdSet<Animal> PolymorphicPossible { get; }
+    }
   }
 }
