--- rd-net/Test.RdFramework/Reflection/data/FieldsNotNullOk.cs
@@ -1,16 +1,0 @@
-﻿using JetBrains.Annotations;
-using JetBrains.Rd.Reflection;
-
-namespace Test.RdFramework.Reflection
-{
-  [RdModel]
-  public sealed class FieldsNotNullOk : RdReflectionBindableBase
-  {
-    public FieldsNotNullOk()
-    {
-    }
-
-    [NotNull] public string FieldOne;
-    public string FieldTwo; // NotNull by default
-  }
-}
