--- rd-net/Test.RdFramework/Reflection/data/Mammal.cs
@@ -1,7 +1,0 @@
-﻿using JetBrains.Rd.Reflection;
-
-namespace Test.RdFramework.Reflection
-{
-  [RdModel]
-  public class Mammal : Animal {}
-}
