--- rd-net/Test.RdFramework/Reflection/data/MyEnum.cs
@@ -1,7 +1,0 @@
-﻿namespace Test.RdFramework.Reflection
-{
-  public enum MyEnum
-  {
-    First, Second
-  }
-}
