--- rd-net/Test.RdFramework/Reflection/data/NestedModel.cs
@@ -1,16 +1,0 @@
-﻿using JetBrains.Collections.Viewable;
-using JetBrains.Rd.Impl;
-using JetBrains.Rd.Reflection;
-
-namespace Test.RdFramework.Reflection
-{
-  [RdExt]
-  public sealed class NestedModel : RdExtReflectionBindableBase
-  {
-    public RdProperty<string> SomeProperty { get; }
-    public IViewableProperty<string> IPropertyAlsoFine { get; }
-
-    public IViewableProperty<PropertiesNotNullOk> PropModel { get; }
-    public RdSet<Animal> PolymorphicPossible { get; }
-  }
-}
