--- rd-net/Test.RdFramework/Reflection/data/NotSealedRdModelData.cs
@@ -1,7 +1,0 @@
-﻿using JetBrains.Rd.Reflection;
-
-namespace Test.RdFramework.Reflection
-{
-  [RdExt]
-  public class NotSealedRdModelData : RdExtReflectionBindableBase { }
-}
